+++ api/web/src/components/CloudTAK/Menu/Overlays.vue
@@ -1,178 +1,170 @@
 <template>
-    <div>
-        <OverlayLayers
-            v-if='isEditing'
-            :overlay='isEditing'
-            @close='isEditing = false'
-        />
+    <MenuTemplate name='Overlays'>
+        <template #buttons>
+            <IconPencil
+                v-if='isDraggable === false'
+                v-tooltip='"Edit Order"'
+                class='cursor-pointer'
+                :size='32'
+                :stroke='1'
+                @click='isDraggable = true'
+            />
+            <IconPencilCheck
+                v-else-if='isDraggable === true'
+                v-tooltip='"Save Order"'
+                class='cursor-pointer'
+                :size='32'
+                :stroke='1'
+                @click='isDraggable = false'
+            />
 
-        <MenuTemplate
-            v-else
-            name='Overlays'
-        >
-            <template #buttons>
-                <IconPencil
-                    v-if='isDraggable === false'
-                    v-tooltip='"Edit Order"'
-                    class='cursor-pointer'
-                    :size='32'
-                    :stroke='1'
-                    @click='isDraggable = true'
-                />
-                <IconPencilCheck
-                    v-else-if='isDraggable === true'
-                    v-tooltip='"Save Order"'
-                    class='cursor-pointer'
-                    :size='32'
-                    :stroke='1'
-                    @click='isDraggable = false'
-                />
+            <IconPlus
+                v-if='!isDraggable'
+                v-tooltip='"Add Overlay"'
+                class='cursor-pointer'
+                :size='32'
+                :stroke='1'
+                @click='$router.push("/menu/datas")'
+            />
+        </template>
+        <template #default>
+            <TablerLoading v-if='loading || !isLoaded' />
+            <template v-else>
+                <div ref='sortable'>
+                    <div
+                        v-for='overlay in overlays'
+                        :id='overlay.id'
+                        :key='overlay.id'
+                        class='col-lg py-2'
+                    >
+                        <div class='py-2 px-3'>
+                            <div class='col-12 d-flex align-items-center'>
+                                <IconGripVertical
+                                    v-if='isDraggable'
+                                    v-tooltip='"Draw to reorder"'
+                                    class='drag-handle cursor-move'
+                                    :size='20'
+                                    :stroke='1'
+                                />
 
-                <IconPlus
-                    v-if='!isDraggable'
-                    v-tooltip='"Add Overlay"'
-                    class='cursor-pointer'
-                    :size='32'
-                    :stroke='1'
-                    @click='$router.push("/menu/datas")'
-                />
-            </template>
-            <template #default>
-                <TablerLoading v-if='loading || !isLoaded' />
-                <template v-else>
-                    <div ref='sortable'>
-                        <div
-                            v-for='overlay in overlays'
-                            :id='overlay.id'
-                            :key='overlay.id'
-                            class='col-lg py-2'
-                        >
-                            <div class='py-2 px-3'>
-                                <div class='col-12 d-flex align-items-center'>
-                                    <IconGripVertical
-                                        v-if='isDraggable'
-                                        v-tooltip='"Draw to reorder"'
-                                        class='drag-handle cursor-move'
+                                <IconChevronRight
+                                    v-if='!isDraggable && !opened.includes(overlay.id)'
+                                    :size='20'
+                                    :stroke='1'
+                                    class='cursor-pointer'
+                                    @click='opened.push(overlay.id)'
+                                />
+                                <IconChevronDown
+                                    v-else-if='!isDraggable'
+                                    :size='20'
+                                    :stroke='1'
+                                    class='cursor-pointer'
+                                    @click='opened.splice(opened.indexOf(overlay.id), 1)'
+                                />
+
+                                <span class='mx-2'>
+                                    <IconMap
+                                        v-if='overlay.type === "raster"'
+                                        v-tooltip='"Raster"'
                                         :size='20'
                                         :stroke='1'
                                     />
+                                    <IconVector
+                                        v-else
+                                        v-tooltip='"Vector"'
+                                        :size='20'
+                                        :stroke='1'
+                                    />
+                                </span>
 
-                                    <IconChevronRight
-                                        v-if='!isDraggable && !opened.includes(overlay.id)'
+                                <span
+                                    class='mx-2 user-select-none'
+                                    :class='{
+                                        "cursor-pointer": ["data", "profile"].includes(overlay.mode) && overlay.type === "vector"
+                                    }'
+                                    v-text='overlay.name'
+                                />
+
+                                <div class='ms-auto btn-list'>
+                                    <IconMaximize
+                                        v-if='getSource(overlay).bounds'
+                                        v-tooltip='"Zoom To Overlay"'
                                         :size='20'
                                         :stroke='1'
                                         class='cursor-pointer'
-                                        @click='opened.push(overlay.id)'
+                                        @click.stop.prevent='zoomTo(overlay)'
                                     />
-                                    <IconChevronDown
-                                        v-else-if='!isDraggable'
+                                    <TablerDelete
+                                        v-if='opened.includes(overlay.id) && ["mission", "data", "profile"].includes(overlay.mode)'
+                                        :key='overlay.id'
+                                        v-tooltip='"Delete Overlay"'
+                                        :size='20'
+                                        displaytype='icon'
+                                        @delete='removeOverlay(overlay)'
+                                    />
+
+                                    <IconEye
+                                        v-if='overlay.visible'
+                                        v-tooltip='"Hide Layer"'
                                         :size='20'
                                         :stroke='1'
                                         class='cursor-pointer'
-                                        @click='opened.splice(opened.indexOf(overlay.id), 1)'
+                                        @click.stop.prevent='overlay.update({ visible: !overlay.visible })'
                                     />
-
-                                    <span class='mx-2'>
-                                        <IconMap
-                                            v-if='overlay.type === "raster"'
-                                            v-tooltip='"Raster"'
-                                            :size='20'
-                                            :stroke='1'
-                                        />
-                                        <IconVector
-                                            v-else
-                                            v-tooltip='"Vector"'
-                                            :size='20'
-                                            :stroke='1'
-                                        />
-                                    </span>
-
-                                    <span
-                                        class='mx-2 user-select-none'
-                                        :class='{
-                                            "cursor-pointer": ["data", "profile"].includes(overlay.mode) && overlay.type === "vector"
-                                        }'
-                                        @click='editor(overlay)'
-                                        v-text='overlay.name'
+                                    <IconEyeOff
+                                        v-else
+                                        v-tooltip='"Show Layer"'
+                                        :size='20'
+                                        :stroke='1'
+                                        class='cursor-pointer'
+                                        @click.stop.prevent='overlay.update({ visible: !overlay.visible })'
                                     />
-
-                                    <div class='ms-auto btn-list'>
-                                        <IconMaximize
-                                            v-if='getSource(overlay).bounds'
-                                            v-tooltip='"Zoom To Overlay"'
-                                            :size='20'
-                                            :stroke='1'
-                                            class='cursor-pointer'
-                                            @click.stop.prevent='zoomTo(overlay)'
-                                        />
-                                        <TablerDelete
-                                            v-if='opened.includes(overlay.id) && ["mission", "data", "profile"].includes(overlay.mode)'
-                                            :key='overlay.id'
-                                            v-tooltip='"Delete Overlay"'
-                                            :size='20'
-                                            displaytype='icon'
-                                            @delete='removeOverlay(overlay)'
-                                        />
-
-                                        <IconEye
-                                            v-if='overlay.visible'
-                                            v-tooltip='"Hide Layer"'
-                                            :size='20'
-                                            :stroke='1'
-                                            class='cursor-pointer'
-                                            @click.stop.prevent='overlay.update({ visible: !overlay.visible })'
-                                        />
-                                        <IconEyeOff
-                                            v-else
-                                            v-tooltip='"Show Layer"'
-                                            :size='20'
-                                            :stroke='1'
-                                            class='cursor-pointer'
-                                            @click.stop.prevent='overlay.update({ visible: !overlay.visible })'
-                                        />
-                                    </div>
                                 </div>
                             </div>
-
-                            <template v-if='!isDraggable'>
-                                <div
-                                    v-if='overlay.type === "raster" && opened.includes(overlay.id)'
-                                    class='col-12'
-                                    style='margin-left: 30px; padding-right: 40px;'
-                                >
-                                    <TablerRange
-                                        v-model='overlay.opacity'
-                                        label='Opacity'
-                                        :min='0'
-                                        :max='1'
-                                        :step='0.1'
-                                        @change='overlay.update({
-                                            opacity: overlay.opacity
-                                        })'
-                                    />
-                                </div>
-                                <TreeCots
-                                    v-else-if='overlay.type === "geojson" && overlay.id === -1 && opened.includes(overlay.id)'
-                                    :element='overlay'
-                                />
-                            </template>
                         </div>
+
+                        <template v-if='!isDraggable && opened.includes(overlay.id)'>
+                            <div
+                                v-if='overlay.type === "raster"'
+                                class='col-12'
+                                style='margin-left: 30px; padding-right: 40px;'
+                            >
+                                <TablerRange
+                                    v-model='overlay.opacity'
+                                    label='Opacity'
+                                    :min='0'
+                                    :max='1'
+                                    :step='0.1'
+                                    @change='overlay.update({
+                                        opacity: overlay.opacity
+                                    })'
+                                />
+                            </div>
+                            <TreeCots
+                                v-if='overlay.type === "geojson" && overlay.id === -1'
+                                :element='overlay'
+                            />
+                            <TreeVector
+                                v-if='overlay.type === "vector"'
+                                :overlay='overlay'
+                            />
+                        </template>
                     </div>
-                </template>
+                </div>
             </template>
-        </MenuTemplate>
-    </div>
+        </template>
+    </MenuTemplate>
 </template>
 
 <script>
 import MenuTemplate from '../util/MenuTemplate.vue';
-import OverlayLayers from './Overlays/Layers.vue';
 import {
     TablerDelete,
     TablerLoading,
     TablerRange
 } from '@tak-ps/vue-tabler';
 import TreeCots from './Overlays/TreeCots.vue';
+import TreeVector from './Overlays/TreeVector.vue';
 import {
     IconGripVertical,
     IconChevronRight,
@@ -199,7 +191,6 @@ export default {
         return {
             err: false,
             isDraggable: false,
-            isEditing: false,
             loading: false,
             opened: [],
             treeState: {
@@ -261,11 +252,6 @@ export default {
             await mapStore.removeOverlay(overlay);
             this.loading = false;
         },
-        editor: function(overlay) {
-            if (["data", "profile"].includes(overlay.mode) && overlay.type === "vector") {
-                this.isEditing = overlay;
-            }
-        },
         getSource: function(overlay) {
             return mapStore.map.getSource(String(overlay.id))
         },
@@ -278,7 +264,7 @@ export default {
     components: {
         MenuTemplate,
         TreeCots,
-        OverlayLayers,
+        TreeVector,
         TablerRange,
         TablerLoading,
         TablerDelete,
